thay đổi tên Thời Gian -> Thời Gian Game

+++ My_Game/Camera.cpp
@@ -4,9 +4,9 @@ RECT Camera::NhanGioiHanHCN()
 {
 	RECT lHCN;
 
-	lHCN.left = mViTri.x - mChieuRong / 2;
+	lHCN.left = int(mViTri.x - mChieuRong / 2);
 	lHCN.right = lHCN.left + mChieuRong;
-	lHCN.top = mViTri.y - mChieuCao / 2;
+	lHCN.top = int(mViTri.y - mChieuCao / 2);
 	lHCN.bottom = lHCN.top + mChieuCao;
 
 	return lHCN;
